fix the issue of dircolors lready installed"

+++ install.sh
@@ -43,6 +43,10 @@ if [[ ! -e ${HOME}/.dircolors ]]; then
 	read -p "This will enable dircolors. Are you sure? [Y/N]" -n 1;
 	printf "";
 	if [[ $REPLY =~ ^[Yy]$ ]]; then
-		ln -s "$SOURCE_LOCATION/dotfiles/dircolors/LS_COLORS" "${HOME}/.dircolors"
+        if [[ -f ${HOME}/.dircolors ]]; then
+          echo "dircolors installation dirctory found, skipping .."
+        else
+            ln -s "$SOURCE_LOCATION/dotfiles/dircolors/LS_COLORS" "${HOME}/.dircolors"
+        fi
 	fi;
 fi
